requirements.txt
fastapi
uvicorn
sentence-transformers
pinecone
python-dotenv
openai
pydantic
boto3
python-jose[cryptography]
bcrypt
python-multipart
